feat(reactive ws): adding services, controllers and mappers+fix

+++ 02 - Reactive Microservices with Spring WebFlux/microservices/orders-service/src/main/java/com/reactive/orders/ws/io/entity/PurchaseOrder.java
@@ -15,10 +15,10 @@ public class PurchaseOrder {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String productId;
-    private Integer userId;
-    private Integer amount;
+    private Long userId;
+    private Double amount;
     private OrderStatus status;
 
 }
